refractoring cache
